Map: fix elevation mapping
along with return type to make it work

#1385

+++ gemrb/core/Scriptable/Actor.cpp
@@ -8442,7 +8442,7 @@ bool Actor::HasBodyHeat() const
 	return true;
 }
 
-uint8_t Actor::GetElevation() const
+int Actor::GetElevation() const
 {
 	return area ? area->GetHeight(Pos) : 0;
 }
